requirements.txt
streamlit
streamlit-option-menu
pandas
openai
pyyaml
requests